Build a Question migration to use as a model
I create a Question migration which will be used in the application. A
user can ask questions in the app. Users can also comment on the
questions. This PR creates the views for the Question model.

For example. A user can create a question, by submitting a title and the
body for the question. Other users can then create comments. I decided
to use Question since I also had Post.