Change file from scss to css for tailwind config
When deploying to Heroku I was getting the following error message.

```
remote:            @import "tailwindcss/base";
remote:            ^
remote:                  File to import not found or unreadable: tailwindcss/base.
```

It seems the reason has to do with tailwind use of post-cssimport the
application file being a SASS file.

```
/* Using postcss-import */
@import "tailwindcss/base";
@import "tailwindcss/components";
@import "tailwindcss/utilities";
@import "./custom-utilities.css";
/* Using Sass or Less */
@tailwind base;
@tailwind components;
@tailwind utilities;
@import "./custom-utilities";
```

This should allow for webpack to find the base dependency when deploying
to Heroku.